Change request rate limit type to int. Fix snapshot loading.

+++ yt/server/security_server/request_tracker.h
@@ -43,7 +43,7 @@ public:
 
     void SetUserRequestRateLimit(
         TUser* user,
-        double limit);
+        int limit);
 
     void ReconfigureUserRequestRateThrottler(TUser* user);
